--- CHANGELOG.txt
@@ -1,5 +1,0 @@
-==========
-
-1.1.9 (11/02/2021)
--------------------
-- First Release
